### 2021-11-12 #3912 @whedon
[PRE REVIEW]: WaveletsExt.jl: Extending the boundaries of wavelets in Julia
**Submitting author:** @ShozenD (<a href="http://orcid.org/0000-0001-5575-641X">Shozen Dan</a>)
**Repository:** <a href="https://github.com/UCD4IDS/WaveletsExt.jl" target ="_blank">https://github.com/UCD4IDS/WaveletsExt.jl</a>
**Version:** v0.1.10
