### 2021-11-13 02f3e01
2D ACWT implementation; 2D ACWT tests; 2D transforms manual all completed. ACWT API docs require some work.
