+++ src/mod/acwt_one_level.jl
@@ -1,7 +1,7 @@
 # ========== Single Step Autocorrelation Wavelet Transform ==========
 ## 1D ##
 """
-    acdwt_step(v, j, h, g)
+    acdwt_step(v, d, h, g)
 
 Performs one level of the autocorrelation discrete wavelet transform (acdwt) on the vector
 `v`, which is the j-th level scaling coefficients (Note the 0th level scaling coefficients
@@ -9,7 +9,7 @@ is the raw signal). The vectors `h` and `g` are the detail and scaling filters.
 
 # Arguments
 - `v::AbstractVector{T} where T<:Number`: Vector of coefficients from a node at level `d`.
-- `j::Integer`: Depth level of `v`.
+- `d::Integer`: Depth level of `v`.
 - `h::Vector{S} where S<:Number`: High pass filter.
 - `g::Vector{S} where S<:Number`: Low pass filter.
 
@@ -81,7 +81,7 @@ ACWT.acdwt_step!(w₁, w₂, v, 0, h, g)
 function acdwt_step!(w₁::AbstractVector{T},
                      w₂::AbstractVector{T},
                      v::AbstractVector{T}, 
-                     j::Integer, 
+                     d::Integer, 
                      h::Array{T,1}, 
                      g::Array{T,1}) where {T<:Number, S<:Number}
     # Sanity check
@@ -94,12 +94,12 @@ function acdwt_step!(w₁::AbstractVector{T},
   
     # One step of autocorrelation transform
     for i in 1:N
-        t = i+(1<<j) |> t -> t>N ? mod1(t,N) : t
-        i = mod1(i + (L÷2+1) * 2^j,N) # Need to shift by half the filter size because of periodicity assumption 
+        t = i+(1<<d) |> t -> t>N ? mod1(t,N) : t
+        i = mod1(i + (L÷2+1) * 2^d,N) # Need to shift by half the filter size because of periodicity assumption 
         @inbounds w₁[i] = g[1] * v[t]
         @inbounds w₂[i] = h[1] * v[t]
         for n in 2:L
-            t = t+(1<<j) |> t -> t>N ? mod1(t,N) : t
+            t = t+(1<<d) |> t -> t>N ? mod1(t,N) : t
             @inbounds w₁[i] += g[n] * v[t]
             @inbounds w₂[i] += h[n] * v[t]
         end
@@ -188,4 +188,102 @@ function iacdwt_step!(v::AbstractVector{T},
     for i in eachindex(v)
         @inbounds v[i] = (w₁[i]+w₂[i]) / √2
     end
+end
+
+## 2D ##
+# Forward transform with allocation
+function acdwt_step(v::AbstractMatrix{T}, d::Integer, h::Vector{S}, g::Vector{S}) where
+                   {T<:Number, S<:Number}
+    n, m = size(v)
+    w₁ = Matrix{T}(undef, (n,m))
+    w₂ = Matrix{T}(undef, (n,m))
+    w₃ = Matrix{T}(undef, (n,m))
+    w₄ = Matrix{T}(undef, (n,m))
+    temp = Array{T,3}(undef, (n,m,2))
+    acdwt_step!(w₁, w₂, w₃, w₄, v, d, h, g, temp)
+    return w₁, w₂, w₃, w₄
+end
+# Forward transform without allocation
+function acdwt_step!(w₁::AbstractMatrix{T}, w₂::AbstractMatrix{T},
[…]
