+++ src/mod/ACWT.jl
@@ -58,20 +58,24 @@ acdwt(x, wt, 4) # level 4 decomposition
 
 **See also:** [`acdwt_step`](@ref), [`iacdwt`](@ref), [`acdwt!`](@ref)
 """
-function acdwt(x::AbstractVector{T},
+function acdwt(x::AbstractArray{T},
                wt::OrthoFilter,
                L::Integer = maxtransformlevels(x)) where T<:Number
     # Sanity check
+    @assert 1 ≤ ndims(x) ≤ 2
     @assert L ≤ maxtransformlevels(x) || throw(ArgumentError("Too many transform levels (length(x) < 2^L"))
     @assert L ≥ 1 || throw(ArgumentError("L must be ≥ 1"))
     # Setup
-    n = length(x)
-    xw = Array{T,2}(undef, (n,L+1))
+    sz = size(x)
+    N = ndims(x)+1
+    k = N==2 ? L+1 : 3*L+1
+    xw = Array{T,N}(undef, (sz...,k))
     # Compute transforms
     acdwt!(xw, x, wt, L)
     return xw
 end
 
+# 1D ACDWT without allocation
 @doc raw"""
     acdwt!(xw, x, wt, L)
 
@@ -124,6 +128,34 @@ function acdwt!(xw::AbstractArray{T,2},
     end    
     return xw
 end
+# 2D ACDWT without allocation
+function acdwt!(xw::AbstractArray{T,3},
+                x::AbstractMatrix{T},
+                wt::OrthoFilter,
+                L::Integer = maxtransformlevels(x)) where T<:Number
+    # Sanity check
+    @assert L ≤ maxtransformlevels(x) ||
+        throw(ArgumentError("Too many transform levels"))
+    @assert L ≥ 1 || throw(ArgumentError("L must be ≥ 1"))
+    @assert size(xw,3) == 3*L+1
+
+    # Setup
+    n, m = size(x)
+    g, h = make_acreverseqmfpair(wt)
+    xw[:,:,end] = x
+    temp = Array{T,3}(undef, (n,m,2))
+
+    # ACDWT
+    for d in 0:(L-1)
+        @inbounds v = xw[:,:, 3*(L-d)+1]       # Parent node
+        @inbounds w₁ = @view xw[:,:, 3*(L-d)-2]          # Scaling + Scaling coefficients
+        @inbounds w₂ = @view xw[:,:, 3*(L-d)-1]          # Detail + Scaling coefficients
+        @inbounds w₃ = @view xw[:,:, 3*(L-d)]            # Scaling + Detail coefficients
+        @inbounds w₄ = @view xw[:,:, 3*(L-d)+1]          # Detail + Detail coefficients
+        @inbounds acdwt_step!(w₁, w₂, w₃, w₄, v, d, h, g, temp)
+    end    
+    return xw
+end
 
 ## 2D ##
 """
@@ -133,16 +165,16 @@ Computes the column-wise discrete autocorrelation transform coeficients for 2D s
 
 **See also:** [`vacdwt`](@ref)
 """
-function hacdwt(x::AbstractArray{T,2}, 
-               wt::OrthoFilter, 
-               L::Integer=maxtransformlevels(x,2)) where T<:Number
-    nrow, ncol = size(x)
-    W = Array{T,3}(undef,nrow,L+1,ncol)
-    for i in 1:ncol
-        @inbounds W[:,:,i] = acdwt(x[:,i],wt,L)
-    end
-    return W
-end
+# function hacdwt(x::AbstractArray{T,2}, 
+#                wt::OrthoFilter, 
+#                L::Integer=maxtransformlevels(x,2)) where T<:Number
[…]

[…]
