### 2021-11-13 ef98236
Changes to ACWT and SWT API documentations.
