### 2021-11-13 5f1f79b
Remove trash image file. Update README.md and prepare for next package update.
