+++ .github/workflows/CI.yml
@@ -46,6 +46,8 @@ jobs:
             ${{ runner.os }}-
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
+        env:
+          GKSwstype: "100" # https://discourse.julialang.org/t/generation-of-documentation-fails-qt-qpa-xcb-could-not-connect-to-display/60988
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1 
         with:
