### 2021-11-14 4e09bcd
CompatHelper: add new compat entry for "Images" at version "0.24" for package docs
