### 2021-11-14 8fd6148
CompatHelper: add new compat entry for "TestImages" at version "1" for package docs
