+++ docs/Project.toml
@@ -9,5 +9,6 @@ WaveletsExt = "8f464e1e-25db-479f-b0a5-b7680379e03f"
 [compat]
 Documenter = "0.27.6"
 Plots = "1.21.3"
+TestImages = "1"
 Wavelets = "0.9"
 WaveletsExt = "0.1"
