### 2021-11-14 683a450
Add tests for plot_tfbdry2.
