### 2021-11-14 d2ea3bd
Fix best basis bug to allow compatibility for partial wavelet decompositions. Removed TIWT. Get ready for next patch update.
