+++ src/mod/TIWT.jl
@@ -1,572 +1,0 @@
-module TIWT
-export tidwt,
-       tiwpt,
-       tiwpd,
-       itidwt,
-       itiwpt,
-       itiwpd
-
-using Wavelets
-
-using ..Utils
-
-# ========== Single Step Translation-Invariant Wavelet Transform ==========
-"""
-    tidwt_step(v, d, h, g)
-
-Performs one level of the translation-invariant discrete wavelet transform (TWDWT) on the
-vector `v`, which is the `d`-th level scaling coefficients (Note the 0-th level scaling
-coefficients is the raw signal). The vectors `h` and `g` are the detail and scaling filters.
-
-# Arguments
-- `v::AbstractVector{T} where T<:Number`: Vector of coefficients from a node at level `d`.
-- `d::Integer`: Depth level of `v`.
-- `h::Vector{S} where S<:Number`: High pass filter.
-- `g::Vector{S} where S<:Number`: Low pass filter.
-
-# Returns
-- `w₁::Vector{T}`: Output from the low pass filter.
-- `w₂::Vector{T}`: Output from the high pass filter.
-
-# Examples
-```julia
-using Wavelets, WaveletsExt
-
-# Setup
-v = randn(8)
-wt = wavelet(WT.haar)
-g, h = WT.makereverseqmfpair(wt, true)
-
-# One step of TIDWT
-tidwt_step(v, 0, h, g)
-```
-
-**See also:** [`tidwt_step!`](@ref)
-"""
-function tidwt_step(v::AbstractVector{T}, d::Integer, h::Vector{S}, g::Vector{S}) where
-                   {T<:Number, S<:Number}
-    n = length(v)
-    w₁ = zeros(T, n)
-    w₂ = zeros(T, n)
-
-    tidwt_step!(w₁, w₂, v, d, h, g)
-    return w₁, w₂
-end
-
-"""
-    tidwt_step!(w₁, w₂, v, d, h, g)
-
-Same as `tidwt_step` but without array allocation.
-
-# Arguments
-- `w₁::AbstractVector{T} where T<:Number`: Vector allocation for output from low pass
-  filter.
-- `w₂::AbstractVector{T} where T<:Number`: Vector allocation for output from high pass
-  filter.
-- `v::AbstractVector{T} where T<:Number`: Vector of coefficients from a node at level `d`.
-- `d::Integer`: Depth level of `v`.
-- `h::Vector{S} where S<:Number`: High pass filter.
-- `g::Vector{S} where S<:Number`: Low pass filter.
-
-# Returns
-- `w₁::Vector{T}`: Output from the low pass filter.
-- `w₂::Vector{T}`: Output from the high pass filter.
-
-# Examples
-```julia
-using Wavelets, WaveletsExt
-
-# Setup
-v = randn(8)
-wt = wavelet(WT.haar)
-g, h = WT.makereverseqmfpair(wt, true)
-w₁ = zeros(8)
-w₂ = zeros(8)
-
-# One step of TIDWT
-tidwt_step!(w₁, w₂, v, 0, h, g)
-```
-
-**See also:** [`tidwt_step`](@ref)
-"""
-function tidwt_step!(w₁::AbstractVector{T},
-                     w₂::AbstractVector{T},
-                     v::AbstractVector{T},
-                     d::Integer,
-                     h::Vector{S},
-                     g::Vector{S}) where {T<:Number, S<:Number}
-    # Variable allocations to be used in Wavelets.Transforms.filtdown! 
-    n = length(v)                           # Signal length
-    nc = Utils.nodelength(n, d+1)           # Node length at level d+1
-    filtlen = length(h)                     # Filter length
[…]

[…]
