### 2021-11-14 526d6e1
Fix spelling error in docs/Project.toml
