### 2021-11-14 c1107a5
Removed TIWT from src/WaveletsExt.jl
