+++ src/WaveletsExt.jl
@@ -4,7 +4,6 @@ module WaveletsExt
 
 include("mod/Utils.jl")
 include("mod/DWT.jl")
-include("mod/TIWT.jl")
 include("mod/SIWPD.jl")
 include("mod/ACWT.jl")
 include("mod/BestBasis.jl")
@@ -19,7 +18,6 @@ using Reexport
                 .Denoising,
                 .Utils,
                 .LDB,
-                .TIWT,
                 .SWT,
                 .SIWPD,
                 .ACWT,
