### 2021-11-14 08e1114
Add GKSws settings in test/runtests.jl
