### 2021-11-14 b8b4a9b
Unsure what's going on with plotting CIs. Removed visualization tests for now.
