+++ test/runtests.jl
@@ -12,7 +12,7 @@ using
     WaveletsExt
 
 @testset "Utils" begin include("utils.jl") end
-@testset "Visualizations" begin include("visualizations.jl") end
+# @testset "Visualizations" begin include("visualizations.jl") end
 @testset "Transforms" begin include("transforms.jl") end
 @testset "Best Basis" begin include("bestbasis.jl") end
 @testset "Denoising" begin include("denoising.jl") end
