### 2021-11-14 0ecefd3
Removed plotting test in denoising.
