### 2021-11-26 7a29d48
CompatHelper: bump compat for "Images" to "0.25" for package docs
