+++ docs/Project.toml
@@ -8,7 +8,7 @@ WaveletsExt = "8f464e1e-25db-479f-b0a5-b7680379e03f"
 
 [compat]
 Documenter = "0.27.6"
-Images = "0.24"
+Images = "0.24, 0.25"
 Plots = "1.21.3"
 TestImages = "1"
 Wavelets = "0.9"
