### 2021-11-30 5b5beb1
Split LDB module to multiple files. Energy map functions now compatible with 2D signals.
