+++ src/mod/ldb_measures.jl
@@ -1,0 +1,382 @@
+## DISCRIMINANT MEASURES
+"""
+Discriminant measure for Local Discriminant Basis. Current available subtypes
+are:
+- [`ProbabilityDensityDM`](@ref)
+- [`SignaturesDM`](@ref)
+"""
+abstract type DiscriminantMeasure end
+
+"""
+Discriminant measure for Probability Density and Time Frequency based energy 
+maps. Current available measures are:
+- [`AsymmetricRelativeEntropy`](@ref)
+- [`SymmetricRelativeEntropy`](@ref)
+- [`LpDistance`](@ref)
+- [`HellingerDistance`](@ref)
+"""
+abstract type ProbabilityDensityDM <: DiscriminantMeasure end
+
+"""
+Discriminant measure for Signatures based energy maps. Current available
+measures are:
+- [`EarthMoverDistance`](@ref)
+"""
+abstract type SignaturesDM <: DiscriminantMeasure end
+
+@doc raw"""
+    AsymmetricRelativeEntropy <: ProbabilityDensityDM
+
+Asymmetric Relative Entropy discriminant measure for the Probability Density and
+Time Frequency based energy maps. This measure is also known as cross entropy 
+and Kullback-Leibler divergence.
+
+Equation: ``D(p,q) = \sum p(x) \log \frac{p(x)}{q(x)}``
+"""
+struct AsymmetricRelativeEntropy <: ProbabilityDensityDM end
+
+@doc raw"""
+    SymmetricRelativeEntropy <: ProbabilityDensityDM
+
+Symmetric Relative Entropy discriminant measure for the Probability Density and 
+Time Frequency energy maps. Similar idea to the Asymmetric Relative Entropy, but 
+this aims to make the measure more symmetric.
+
+Equation: Denote the Asymmetric Relative Entropy as ``D_A(p,q)``, then
+
+``D(p,q) = D_A(p,q) + D_A(q,p) = \sum p(x) \log \frac{p(x)}{q(x)} + q(x) \log \frac{q(x)}{p(x)}``
+"""
+struct SymmetricRelativeEntropy <: ProbabilityDensityDM end
+
+@doc raw"""
+    LpDistance <: ProbabilityDensityDM
+
+``\ell^p`` Distance discriminant measure for the Probability Density and Time 
+Frequency based energy maps. The default ``p`` value is set to 2.
+
+Equation: ``W(q,r) = ||q-r||_p^p = \sum_{i=1}^n (q_i - r_i)^p``
+"""
+@with_kw struct LpDistance <: ProbabilityDensityDM 
+    p::Number = 2
+end
+
+@doc raw"""
+    HellingerDistance <: ProbabilityDensityDM
+
+Hellinger Distance discriminant measure for the Probability Density energy 
+map.
+
+Equation: ``H(p,q) = \sum_{i=1}^n (\sqrt{p_i} - \sqrt{q_i})^2``
+"""
+struct HellingerDistance <: ProbabilityDensityDM end
+
+@doc raw"""
+    EarthMoverDistance <: SignaturesDM
+
+Earth Mover Distance discriminant measure for the Signatures energy map.
+
+Equation: 
+``E(P,Q) = \frac{\sum_{k=1}^{m+n+1} |\hat p_k - \hat q_k| (r_{k+1} - r_k)}{w_\Sigma}``
+
+where ``r_1, r_2, \ldots, r_{m+n}`` is the sorted list of ``p_1, \ldots, p_m, 
+q_1, \ldots, q_n`` and ``\hat p_k = \sum_{p_i \leq r_k} w_{p_i}``, 
[…]
