+++ src/mod/ldb_energymap.jl
@@ -1,0 +1,204 @@
+import AverageShiftedHistograms: ash, xy, pdf
+
+## ---------- ENERGY MAPS ----------
+"""
+Energy map for Local Discriminant Basis. Current available types are:
+- [`TimeFrequency`](@ref)
+- [`ProbabilityDensity`](@ref)
+"""
+abstract type EnergyMap end
+
+@doc raw"""
+    TimeFrequency <: EnergyMap
+
+An energy map based on time frequencies, a measure based on the differences of 
+derived quantities from projection ``Z_i``, such as mean class energies or 
+cumulants.
+
+**See also:** [`EnergyMap`](@ref), [`ProbabilityDensity`](@ref),
+    [`Signatures`](@ref)
+"""
+struct TimeFrequency <: EnergyMap end
+
+@doc raw"""
+    ProbabilityDensity <: EnergyMap
+
+An energy map based on probability density, a measure based on the differences 
+among the pdfs of ``Z_i``. Since we do not know the true density functions of
+the coefficients, the PDFs are estimated using the Average Shifted Histogram
+(ASH).
+
+**See also:** [`EnergyMap`](@ref), [`TimeFrequency`](@ref), [`Signatures`](@ref)
+"""
+struct ProbabilityDensity <: EnergyMap end
+
+@doc raw"""
+    Signatures <: EnergyMap
+
+An energy map based on signatures, a measure that uses the Earth Mover's Distance (EMD) to
+compute the discriminating  power of a coordinate. Signatures provide us with a fully
+data-driven representation, which can be efficiently used with EMD. This representation is
+more efficient than a histogram and is able to represent complex data structure with fewer
+samples.
+
+Here, a signature for the coefficients in the ``j``-th level, ``k``-th node, ``l``-th index
+of class ``c`` is defined as
+
+``s_{j,k,l}^{(c)} = \{(\alpha_{i;j,k,l}^{(c)}, w_{i;j,k,l}^{(c)})\}_{i=1}^{N_c}``
+
+where ``\alpha_{i;j,k,l}^{(c)}`` and ``w_{i;j,k,l}^{(c)}`` are the expansion coefficients
+and weights at location ``(j,k,l)`` for signal ``i`` of class ``c`` respectively. Currently,
+the two valid types of weights are `:equal` and `:pdf`.
+
+# Argumemts
+- `weight::Symbol`: Type of weight to be used to compute ``w_{i;j,k,l}^{(c)}``. Available
+    methods are `:equal` and `:pdf`. Default is set to `:equal`.
+
+**See also:** [`EnergyMap`](@ref), [`TimeFrequency`](@ref), [`ProbabilityDensity`](@ref)
+"""
+struct Signatures <: EnergyMap 
+    weight::Symbol
+    # Constructor
+    Signatures(weight=:equal) = weight ∈ [:equal, :pdf] ? new(weight) : 
+        throw(ValueError("Invalid weight type. Valid weight types are :equal and :pdf."))
+end
+
+"""
+    energy_map(Xw, y, method)
+
+Returns the Time Frequency Energy map or the Probability Density Energy map
+depending on the input `method` (`TimeFrequency()` or `ProbabilityDensity()`).
+
+**See also:** [`EnergyMap`](@ref). [`TimeFrequency`](@ref), 
+    [`ProbabilityDensity`](@ref)
+"""
+function energy_map(Xw::AbstractArray{S}, y::AbstractVector{T}, method::TimeFrequency) where 
+                   {S<:AbstractFloat, T}
+    # --- Sanity check ---
+    N = ndims(Xw)
+    c = unique(y)       # unique classes
+    nc = length(c)      # number of classes
[…]

[…]
