+++ src/mod/LDB.jl
@@ -33,13 +33,13 @@ export
 
 using LinearAlgebra,
       Wavelets,
-      Parameters,
-      Statistics,
-      StatsBase
+      Statistics
+import Parameters: @with_kw
+import StatsBase: mad
 
 using ..Utils,
-      ..DWT,
-      ..BestBasis
+      ..DWT
+import ..BestBasis: bestbasis_treeselection
 
 include("ldb_energymap.jl")
 include("ldb_measures.jl")
@@ -48,115 +48,116 @@ include("ldb_measures.jl")
 """
     LocalDiscriminantBasis
 
-Class type for the Local Discriminant Basis (LDB), a feature selection algorithm
-developed by N. Saito and R. Coifman in "Local Discriminant Bases and Their
-Applications" in the Journal of Mathematical Imaging and Vision, Vol 5, 337-358
-(1995). This struct contains the following field values: 
+Class type for the Local Discriminant Basis (LDB), a feature selection algorithm developed
+by N. Saito and R. Coifman in "Local Discriminant Bases and Their Applications" in the
+Journal of Mathematical Imaging and Vision, Vol 5, 337-358 (1995). This struct contains the
+following field values: 
 
 # Parameters and Attributes:
-- `wt::DiscreteWavelet`: a discrete wavelet for transform purposes
-- `max_dec_level::Union{Integer, Nothing}`: max level of wavelet packet
-    decomposition to be computed.
-- `dm::DiscriminantMeasure`: the discriminant measure for the LDB algorithm. 
-    Supported measures are the `AsymmetricRelativeEntropy()`, `LpDistance()`,
-    `SymmetricRelativeEntropy()`, and `HellingerDistance()`
-- `en::EnergyMap`: the type of energy map used. Supported maps are 
-    `TimeFrequency()`, `ProbabilityDensity()`, and `Signatures()`.
-- `dp::DiscriminantPower()`: the measure of discriminant power among expansion
-    coefficients. Supported measures are `BasisDiscriminantMeasure()`,
-    `FishersClassSeparability()`, and `RobustFishersClassSeparability()`. 
-- `top_k::Union{Integer, Nothing}`: the top-k coefficients used in each node to 
-    determine the discriminant measure.
-- `n_features::Union{Integer, Nothing}`: the dimension of output after 
+- `wt::DiscreteWavelet`: (Default: `wavelet(WT.haar)`) A discrete wavelet for transform
+  purposes.
+- `max_dec_level::Union{Integer, Nothing}`: (Default: `nothing`) Max level of wavelet packet
+  decomposition to be computed.
+- `dm::DiscriminantMeasure`: (Default: `AsymmetricRelativeEntropy()`) the discriminant
+    measure for the LDB algorithm. Supported measures are the `AsymmetricRelativeEntropy()`,
+    `LpDistance()`, `SymmetricRelativeEntropy()`, and `HellingerDistance()`.
+- `en::EnergyMap`: (Default: `TimeFrequency()`) the type of energy map used. Supported maps
+    are `TimeFrequency()`, `ProbabilityDensity()`, and `Signatures()`.
+- `dp::DiscriminantPower()`: (Default: `BasisDiscriminantMeasure()`) the measure of
+    discriminant power among expansion coefficients. Supported measures are
+    `BasisDiscriminantMeasure()`, `FishersClassSeparability()`, and
+    `RobustFishersClassSeparability()`. 
[…]
