+++ Project.toml
@@ -5,6 +5,7 @@ version = "0.1.12"
 
 [deps]
 AverageShiftedHistograms = "77b51b56-6f8f-5c3a-9cb4-d71f9594ea6e"
+Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 ImageQualityIndexes = "2996bd0c-7a13-11e9-2da2-2f5ce47296a9"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"

[…]
