### 2021-12-02 ce7df3c
Update Julia testing versions on Project.toml and CI file.
