+++ .github/workflows/CI.yml
@@ -20,8 +20,8 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.5'
           - '1.6'
+          - '1.7'
           - 'nightly'
         os:
           - ubuntu-latest
