+++ Project.toml
@@ -1,7 +1,7 @@
 name = "WaveletsExt"
 uuid = "8f464e1e-25db-479f-b0a5-b7680379e03f"
 authors = ["Zeng Fung Liew <lzfliew@ucdavis.edu>", "Shozen Dan <shodan@ucdavis.edu>"]
-version = "0.1.12"
+version = "0.1.13"
 
 [deps]
 AverageShiftedHistograms = "77b51b56-6f8f-5c3a-9cb4-d71f9594ea6e"
@@ -26,4 +26,4 @@ Plots = "1.20"
 Reexport = "1.1"
 StatsBase = "0.33.9"
 Wavelets = "0.9.3"
-julia = "1.5, 1.6"
+julia = "1.6, 1.7"
