### 2021-12-02 09180c3
Change symbols in denoise and denoiseall functions from :acwt to :acdwt, and :acwpt and :acwpd. Run visualizations CI tests again.
