### 2021-12-02 922a0b5
Update documentations for energy map, discriminant measures, and discriminant powers functions. Documentation updates for fit!, transform etc. required.
