+++ src/mod/ldb_measures.jl
@@ -29,25 +29,33 @@ abstract type SignaturesDM <: DiscriminantMeasure end
 @doc raw"""
     AsymmetricRelativeEntropy <: ProbabilityDensityDM
 
-Asymmetric Relative Entropy discriminant measure for the Probability Density and
-Time Frequency based energy maps. This measure is also known as cross entropy 
-and Kullback-Leibler divergence.
+Asymmetric Relative Entropy discriminant measure for the Probability Density and Time
+Frequency based energy maps. This measure is also known as cross entropy and
+Kullback-Leibler divergence.
 
-Equation: ``D(p,q) = \sum p(x) \log \frac{p(x)}{q(x)}``
+Equation: 
+
+``D(p,q) = \sum p(x) \log \frac{p(x)}{q(x)}`` 
+
+where ``\int_{-\infty}^{\infty} p(t) dt = \int_{-\infty}^{\infty} q(t) dt = 1`` (``p(t)``
+and ``q(t)`` are probability density functions.)
 """
 struct AsymmetricRelativeEntropy <: ProbabilityDensityDM end
 
 @doc raw"""
     SymmetricRelativeEntropy <: ProbabilityDensityDM
 
-Symmetric Relative Entropy discriminant measure for the Probability Density and 
-Time Frequency energy maps. Similar idea to the Asymmetric Relative Entropy, but 
-this aims to make the measure more symmetric.
+Symmetric Relative Entropy discriminant measure for the Probability Density and Time
+Frequency energy maps. Similar idea to the Asymmetric Relative Entropy, but this aims to
+make the measure more symmetric.
 
 Equation: Denote the Asymmetric Relative Entropy as ``D_A(p,q)``, then
 
 ``D(p,q) = D_A(p,q) + D_A(q,p) = \sum p(x) \log \frac{p(x)}{q(x)} + q(x) \log
 \frac{q(x)}{p(x)}``
+
+where ``\int_{-\infty}^{\infty} p(t) dt = \int_{-\infty}^{\infty} q(t) dt = 1`` (``p(t)``
+and ``q(t)`` are probability density functions.)
 
 **See also:** [`AsymmetricRelativeEntropy`](@ref)
 """
@@ -59,7 +67,9 @@ struct SymmetricRelativeEntropy <: ProbabilityDensityDM end
 ``\ell^p`` Distance discriminant measure for the Probability Density and Time 
 Frequency based energy maps. The default ``p`` value is set to 2.
 
-Equation: ``W(q,r) = ||q-r||_p^p = \sum_{i=1}^n (q_i - r_i)^p``
+Equation: 
+
+``W(q,r) = ||q-r||_p^p = \sum_{i=1}^n (q_i - r_i)^p``
 """
 @with_kw struct LpDistance <: ProbabilityDensityDM 
     p::Number = 2
@@ -71,7 +81,9 @@ end
 Hellinger Distance discriminant measure for the Probability Density energy 
 map.
 
-Equation: ``H(p,q) = \sum_{i=1}^n (\sqrt{p_i} - \sqrt{q_i})^2``
+Equation: 
+
+``H(p,q) = \sum_{i=1}^n (\sqrt{p_i} - \sqrt{q_i})^2``
 """
 struct HellingerDistance <: ProbabilityDensityDM end
 
@@ -80,19 +92,49 @@ struct HellingerDistance <: ProbabilityDensityDM end
 
 Earth Mover Distance discriminant measure for the Signatures energy map.
 
-Equation: ``E(P,Q) = \frac{\sum_{k=1}^{m+n+1} |\hat p_k - \hat q_k| (r_{k+1} -
-r_k)}{w_\Sigma}``
[…]
