+++ docs/src/api/ldb.md
@@ -4,14 +4,45 @@
 Modules = [LDB]
 ```
 
-## Public API
-```@autodocs
-Modules = [LDB]
-Private = false
+## Energy Maps
+```@docs
+LDB.EnergyMap
+LDB.TimeFrequency
+LDB.ProbabilityDensity
+LDB.Signatures
+LDB.energy_map
 ```
 
-# Private API
-```@autodocs
-Modules = [LDB]
-Public = false
+## Discriminant Measures
+```@docs
+LDB.DiscriminantMeasure
+LDB.ProbabilityDensityDM
+LDB.SignaturesDM
+LDB.AsymmetricRelativeEntropy
+LDB.SymmetricRelativeEntropy
+LDB.HellingerDistance
+LDB.LpDistance
+LDB.EarthMoverDistance
+LDB.discriminant_measure
+LDB.pairwise_discriminant_measure
+```
+
+## Computation of Discriminant Powers
+```@docs
+LDB.DiscriminantPower
+LDB.BasisDiscriminantMeasure
+LDB.FishersClassSeparability
+LDB.RobustFishersClassSeparability
+LDB.discriminant_power
+```
+
+## Feature Extraction and Transformation
+```@docs
+LDB.LocalDiscriminantBasis
+LDB.fit!
+LDB.fitdec!
+LDB.fit_transform
+LDB.transform
+LDB.inverse_transform
+LDB.change_nfeatures
 ```

[…]
