+++ src/mod/LDB.jl
@@ -1,5 +1,4 @@
 module LDB
-using Base: AbstractFloat
 export 
     # energy map
     EnergyMap,
@@ -73,9 +72,9 @@ following field values:
 - `n_features::Union{Integer, Nothing}`: (Default: `nothing`) the dimension of output after
     undergoing feature selection and transformation.
 - `sz::Union{Vector{T} where T<:Integer, Nothing}`: (Default: `nothing`) Size of signal
-- `Γ::Union{AbstractArray{<:AbstractFloat}, AbstractArray{NamedTuple{(:coef, :weight),
-  Tuple{S1, S2}}} where {S1<:Array{T} where T<:AbstractFloat, S2<:Union{T, Array{T}} where
-  T<:AbstractFloat}, Nothing}`: (Default: `nothing`) computed energy map
+- `Γ::Union{AbstractArray{T} where T<:AbstractFloat, AbstractArray{NamedTuple{(:coef,
+  :weight), Tuple{S₁, S₂}}} where {S₁<:Array{T} where T<:AbstractFloat, S2<:Union{T,
+  Array{T}} where T<:AbstractFloat}, Nothing}`: (Default: `nothing`) computed energy map
 - `DM::Union{AbstractArray{<:AbstractFloat}, Nothing}`: (Default: `nothing`) computed
   discriminant measure
 - `cost::Union{AbstractVector{<:AbstractFloat}, Nothing}`: (Default: `nothing`) computed
@@ -113,8 +112,26 @@ end
 """
     fit!(f, X, y)
 
-Fits the Local Discriminant Basis feature selection algorithm `f` onto the signals `X` (or
-the decomposed signals `Xw`) with labels `y`.
+Fits the Local Discriminant Basis feature selection algorithm `f` onto the signals `X` with
+labels `y`.
+
+# Arguments
+- `f::LocalDiscriminantBasis`: Local discriminant basis object.
+- `X::AbstractArray{S} where S<:AbstractFloat`: Group of signals of the same size, arranged
+  in the 2nd (1D signals) or 3rd (2D signals) dimension.
+- `y::AbstractVector{T} where T`: Labels corresponding to each signal in X.
+
+# Returns
+- `f::LocalDiscriminantBasis`: The same LDB object `f` with updated fields.
+
+# Examples
+```julia
+using Wavelets, WaveletsExt
+
+X, y = generateclassdata(ClassData(:tri, 5, 5, 5))
+f = LocalDiscriminantBasis()
+fit!(f, X, y)
+```
 
 **See also:** [`LocalDiscriminantBasis`](@ref), [`fit_transform`](@ref),
     [`transform`](@ref), [`inverse_transform`](@ref), [`change_nfeatures`](@ref)
@@ -140,6 +157,31 @@ end
 
 """
     fitdec!(f, Xw, y)
+
+Fits the Local Discriminant Basis feature selection algorithm `f` onto the decomposed
+signals `Xw` with labels `y`.
+
+# Arguments
+- `f::LocalDiscriminantBasis`: Local discriminant basis object.
+- `Xw::AbstractArray{S} where S<:AbstractFloat`: Group of decomposed signals of the same
+  size, arranged in the 3rd (1D signals) or 4th (2D signals) dimension.
+- `y::AbstractVector{T} where T`: Labels corresponding to each signal in X.
+
+# Returns
+- `f::LocalDiscriminantBasis`: The same LDB object `f` with updated fields.
+
+# Examples
+```julia
+using Wavelets, WaveletsExt
+
+X, y = generateclassdata(ClassData(:tri, 5, 5, 5))
+f = LocalDiscriminantBasis()
[…]
