+++ docs/src/api/bestbasis.md
@@ -4,8 +4,7 @@
 Modules = [BestBasis]
 ```
 
-## Public API
-### Cost functions and computations
+## Cost functions and computations
 ```@docs
 BestBasis.CostFunction
 BestBasis.LSDBCost
@@ -21,7 +20,14 @@ BestBasis.tree_costs
 BestBasis.tree_costs(::AbstractMatrix{T}, ::AbstractVector{BitVector}, ::SIBB) where T<:Number
 ```
 
-### Best basis computation
+# Best Basis Tree Selection
+```@docs
+BestBasis.bestbasis_treeselection
+BestBasis.bestbasis_treeselection(::AbstractVector{Tc}, ::AbstractVector{Tt}) where {Tc<:AbstractVector{<:Union{Number,Nothing}}, Tt<:BitVector}
+BestBasis.delete_subtree!
+```
+
+## Best basis computation
 ```@docs
 BestBasis.BestBasisType
 BestBasis.LSDB
@@ -31,12 +37,4 @@ BestBasis.SIBB
 Wavelets.Threshold.bestbasistree
 Wavelets.Threshold.bestbasistree(::AbstractMatrix{T}, ::Integer, ::SIBB) where T<:Number
 BestBasis.bestbasistreeall
-```
-
-# Private API
-```@docs
-BestBasis.bestbasis_treeselection
-BestBasis.bestbasis_treeselection(::AbstractVector{Tc}, ::AbstractVector{Tt}) where {Tc<:AbstractVector{<:Union{Number,Nothing}}, Tt<:BitVector}
-BestBasis.delete_subtree!
-
 ```
