+++ docs/src/api/denoising.md
@@ -4,7 +4,6 @@
 Modules = [Denoising]
 ```
 
-# Public API
 ## Shrinking Types and Constructors
 ```@docs
 Denoising.RelErrorShrink
@@ -25,7 +24,6 @@ Wavelets.Threshold.denoise
 Denoising.denoiseall
 ```
 
-# Private API
 ## Helper Functions for Threshold Determination and Noise Estimation
 ```@docs
 Denoising.surethreshold

[…]
