+++ v0.1.13/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"api/siwpd/#Shift-Invariant-Wavelet-Packet-Decomposition","page":"SIWPD","title":"Shift Invariant Wavelet Packet Decomposition","text":"","category":"section"},{"location":"api/siwpd/","page":"SIWPD","title":"SIWPD","text":"Modules = [SIWPD]","category":"page"},{"location":"api/siwpd/#Public-API","page":"SIWPD","title":"Public API","text":"","category":"section"},{"location":"api/siwpd/","page":"SIWPD","title":"SIWPD","text":"Modules = [SIWPD]\nPrivate = false","category":"page"},{"location":"api/siwpd/#WaveletsExt.SIWPD.makesiwpdtree-Tuple{Integer, Integer, Integer}","page":"SIWPD","title":"WaveletsExt.SIWPD.makesiwpdtree","text":"makesiwpdtree(n, L, d)\n\nReturns the multi-level, multi-depth binary tree corresponding to the Shift- Invariant Wavelet Packet Decomposition. \n\n\n\n\n\n","category":"method"},{"location":"api/siwpd/#WaveletsExt.SIWPD.siwpd-Union{Tuple{T}, Tuple{AbstractVector{T}, OrthoFilter}, Tuple{AbstractVector{T}, OrthoFilter, Integer}, Tuple{AbstractVector{T}, OrthoFilter, Integer, Integer}} where T<:Number","page":"SIWPD","title":"WaveletsExt.SIWPD.siwpd","text":"siwpd(x, wt[, L=maxtransformlevels(x), d=L])\n\nComputes the Shift-Invariant Wavelet Packet Decomposition originally developed by Cohen, Raz & Malah on the vector x using the discrete wavelet filter wt for L levels with depth d.\n\n\n\n\n\n","category":"method"},{"location":"manual/transforms/#transforms_manual","page":"Transforms","title":"Wavelet Transforms","text":"","category":"section"},{"location":"manual/transforms/","page":"Transforms","title":"Transforms","text":"Wavelet transform is a feature extraction process for decomposing signals into high and low frequency segments. Using a pair of orthonormal wavelet psi in L^2(mathbbR), where L^2(mathbbR) is a Hilbert space of square integrable functions, one can compute","category":"page"},{"location":"manual/transforms/","page":"Transforms","title":"Transforms","text":"y_low = g(x) where x is the signal of interest, g is the low pass filter corresponding to psi, and y_low is the output when x passes through g.\ny_high = h(x) where x is the signal of interest, h is the high pass filter corresponding to psi, and y_high is the output when x passes through h.","category":"page"},{"location":"manual/transforms/","page":"Transforms","title":"Transforms","text":"The wavelet transform can be thought of as an improvement over the Fourier transform due to its ability to preserve information in both the time and frequency domains. It has vast applications in fields such as signal analysis and image compression.","category":"page"},{"location":"manual/transforms/","page":"Transforms","title":"Transforms","text":"As an extension to Wavelets.jl, WaveletsExt.jl offers additional (redundant) wavelet transform techniques via autocorrelation wavelet transforms (Beylkin, Saito), stationary wavelet transforms (Nason, Silverman), and shift invariant wavelet transform (Cohen et. al.).","category":"page"},{"location":"manual/transforms/#Wavelet-Transform-Methods","page":"Transforms","title":"Wavelet Transform Methods","text":"","category":"section"},{"location":"manual/transforms/","page":"Transforms","title":"Transforms","text":"There are essentially 3 methods of wavelet transforms: discrete wavelet transforms, wav
[…]
