### 2021-12-04 UCD4IDS/WaveletsExt.jl#39 @lostanlen
LICENSE is MIT but not recognized by GitHub as such
Hello,

It appears that WaveletExt.jl is released under an MIT license; is that right?
https://github.com/UCD4IDS/WaveletsExt.jl/blob/master/LICENSE

It would be good if this file could be converted to Markdown format so that GitHub will recognize it automatically as MIT:
https://github.com/JuliaLang/julia/blob/master/LICENSE.md
