### 2021-12-04 UCD4IDS/WaveletsExt.jl#40 @lostanlen
Necessary vs. recommended dependencies
Quoting from the review checklist of JOSS:

>  Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

I have searched the word `using` in your package and i have found:
* `using Wavelets`
* `using LinearAlgebra`
* `using Statistics`
* `using Plots`
* `using Distributions`
* `using AverageShiftedHistograms`
* `using Parameters`
* `using Documenter`
* `using Test`
* `using ImageQualityIndexes`
* `using Random`

It would be good to clarify which of these packages are necessary/recommended for (1) running the software (2) running the tests (3) compiling the documentation.
