### 2021-12-04 #3937 @lostanlen
> License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

LICENSE is MIT but not recognized by GitHub as such
https://github.com/UCD4IDS/WaveletsExt.jl/issues/39

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Necessary vs. recommended dependencies
https://github.com/UCD4IDS/WaveletsExt.jl/issues/40
