### 2021-12-04 UCD4IDS/WaveletsExt.jl#40 @zengfung
Hi,

where should such clarifications be documented? We currently have 3 separate Project.toml files
- 1 in the `./` directory for running the software
- 1 in the `test/` directory for running tests
- 1 in the `docs/` directory for documentation

Is there anything we need to do further?
