### 2021-12-04 UCD4IDS/WaveletsExt.jl#39 @zengfung
Hi,

WaveletsExt.jl is released under the 3-Clause BSD license, not the MIT license. Will a conversion to Markdown still be necessary? Thanks.
