### 2021-12-04 UCD4IDS/WaveletsExt.jl#40 @lostanlen
My bad. I wasn't aware of this `Project.toml` mechanism. (i only knew about `REQUIRE`)
noted!
