### 2021-12-05 UCD4IDS/WaveletsExt.jl#39 @lostanlen
Sorry, i confused MIT and BSD-3. (which kind of proves my point: it would be good to clarify)

What i'm saying is that GitHub doesn't make the LICENSE type appear on the front page:

Compare with:
