### 2021-12-05 UCD4IDS/WaveletsExt.jl#39 @lostanlen
This allows GitHub to print a nice human-friendly banner above the legal description of the LICENSE:
