### 2021-12-05 UCD4IDS/WaveletsExt.jl#39 @zengfung
Thanks for the pointers. I've changed the LICENSE filetype to Markdown, but I don't see any expected changes as shown in your screenshots. Do I have to rewrite my LICENSE files?
