### 2021-12-16 a65e754
Speed up struct constructors
Change the codes for structs from:
```julia
struct foo
    x::Integer
end
```
into:
```julia
struct foo{T<:Integer}
    x::T
end
```
