### 2021-12-16 beaa967
Revert struct method for `ClassData()`
New method of building structs that was supposedly slightly faster crashes. Reverting to old
method to ensure things run smoothly. Since there isn't a lot of allocations involved, this
is not a big deal and will be ignored at least for now.
