### 2021-12-18 d72aaf3
Speed up SIWPD functions
*Note: This is a rewrite of the previous commit message due to some mistakes in the previous
commit.*

Sped up SIWPD functions using similar structure as DWT and SWT modules. Given a length 32
signal and a Daubechies 4-tap wavelet, the previous functions' benchmark is:
```julia
julia> @benchmark siwpd(x, wt)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  359.400 μs …  11.286 ms  ┊ GC (min … max): 0.00% … 95.35%
 Time  (median):     395.050 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   442.743 μs ± 453.846 μs  ┊ GC (mean ± σ)
[…]
