+++ src/mod/SIWPD.jl
@@ -18,94 +18,102 @@ by Cohen, Raz & Malah on the vector `x` using the discrete wavelet filter `wt`
 for `L` levels with depth `d`.
 """
 function siwpd(x::AbstractVector{T}, wt::OrthoFilter, 
-        L::Integer=maxtransformlevels(x), d::Integer=L) where T<:Number
-    
-    g, h = WT.makereverseqmfpair(wt, true)       
-    si = Vector{T}(undef, length(wt)-1)                 # tmp filter vector              
-    return siwpd(x, wt, h, g, si, L, d)
+               L::S = maxtransformlevels(x), d::S = L) where {T<:AbstractFloat, S<:Integer}
+    # Sanity check
+    n = length(x)
+    @assert 0 ≤ L ≤ maxtransformlevels(n)
+    @assert 1 ≤ d ≤ L
+
+    # Setup
+    g, h = WT.makereverseqmfpair(wt, true)
+    # y = Matrix{T}(undef, (n, gettreelength(1<<(L+1))))
+    y = zeros(T, (n, gettreelength(1<<(L+1))))
+    @inbounds y[:,1] = x
+
+    # Decomposition
+    for i in axes(y,2)
+        lvl = getdepth(i, :binary)
+        len = nodelength(n, lvl)
+        dₗ = (0 ≤ lvl ≤ L-d) ? d : L-lvl
+        siwpd_subtree!(y, h, g, i, dₗ, 0, L=lvl, len=len)
+    end
+    return y
 end
 
-function siwpd(x::AbstractVector{Tx}, wt::OrthoFilter, h::Vector{T}, 
-        g::Vector{T}, si::Vector{T}, L::Integer, d::Integer=L) where 
-        {Tx<:Number, T<:Number}
+#=======================================================================================
+Recursive calls for decomposition while d>0.
+If shift is 0, ie. ss=0, the main root will not be further decomposed even if d>0.
+=======================================================================================#
+function siwpd_subtree!(y::AbstractMatrix{T₁}, h::Vector{T₂}, g::Vector{T₂}, 
+                        i::S, d::S, ss::S; 
+                        L::S = getdepth(i, :binary),
+                        len::S = nodelength(size(y,1), L)) where 
+                       {T₁<:AbstractFloat, T₂<:AbstractFloat, S<:Integer}
+    # Sanity check
+    n, k = size(y)
+    @assert 0 ≤ L ≤ maxtransformlevels(n)
+    @assert 0 ≤ ss < n
+    @assert 0 ≤ d ≤ getdepth(k,:binary)-L
+    @assert iseven(len) || len == 1
 
-    n = length(x)
-    @assert isdyadic(n)
-    @assert 0 <= L <= maxtransformlevels(n)
-    @assert 1 <= d <= L
+    # --- Base case ---
+    d > 0 || return y
 
-    # decomposed array y 
-    y = Array{T,2}(undef, (n,2^(L+1)-1))
-    y[:,1] = x
-
-    for node in axes(y,2)
-        level = floor(Int, log2(node)) # current node level (starts from 0)
-        nodelen = nodelength(n, level)
-        # decomposition depth of current node
-        dₗ = (0 <= level <= L-d) ? d : L-level  
-        decompose!(y, wt, h, g, si, node, nodelen, 
-            level, dₗ, 0, zeroshiftdecomposition=false)
+    # --- One level of decomposition ---
+    v_start = ss*len+1                                      # Start index for parent node
+    v_end = (ss+1)*len                                      # End index for parent node
+    @inbounds v = @view y[v_start:v_end, i]                 # Parent node
+    w_start = ss*(len÷2)+1                                  # Start index for child node
[…]
