### 2021-12-18 d456167
Update generation of triangular signals
Previous method of generation was buggy and did not produce the right signals.
