### 2021-12-19 UCD4IDS/WaveletsExt.jl#39 @lostanlen
OK  many thanks!
