### 2021-12-19 #3937 @lostanlen
both issues i raised two weeks ago are now closed (39 and 40)
