+++ src/mod/Visualizations.jl
@@ -1,7 +1,9 @@
 module Visualizations
 export 
     plot_tfbdry,
+    plot_tfbdry!,
     plot_tfbdry2,
+    plot_tfbdry2!,
     wiggle,
     wiggle!
 
@@ -76,21 +78,22 @@ function treenodes_matrix(x::BitVector)
 end
 
 """
-    plot_tfbdry(tree[; start, nd_col, ln_col, bg_col])
+    plot_tfbdry([plt], tree[; start, nd_col, ln_col, bg_col])
 
 Given a tree, output a visual representation of the leaf nodes, user will have the option to
 start the node count of each level with 0 or 1.
 
 # Arguments
+- `plt::Plots.Plot`: Plot object.
 - `tree::BitVector`: Tree for plotting the leaf nodes. Comes in the form of a `BitVector`.
 - `depth::Integer`: (Default: `log2(length(tree)+1)-1 |> Int`) Maximum depth to be
   displayed.
 
 # Keyword Arguments
 - `start::Integer`: (Default: `0`) Whether to zero-index or one-index the root of the tree.
-- `nd_col::Symbol`: (Default: `:white`) Color of the leaf nodes.
-- `ln_col::Symbol`: (Default: `:white`) Color of lines in plot.
-- `bg_col::Symbol`: (Default: `:black`) Color of background.
+- `node_color::Symbol`: (Default: `:black`) Color of the leaf nodes.
+- `line_color::Symbol`: (Default: `:black`) Color of lines in plot.
+- `background_color::Symbol`: (Default: `:white`) Color of background.
 
 # Returns
 `::Plots.Plot`: Plot object with the visual representation of the leaf nodes.
@@ -106,15 +109,50 @@ tree = maketree(128, 7, :dwt)
 plot_tfbdry(tree)
 ```
 
-**See also:** [`plot_tfbdry2`](@ref)
+**See also:** [`plot_tfbdry!`](@ref), [`plot_tfbdry2`](@ref)
 """
-function plot_tfbdry(tree::BitVector,
-                     depth::Integer = log2(length(tree)+1)-1 |> Int;
-                     start::Integer = 0, 
-                     nd_col::Symbol = :white,
-                     ln_col::Symbol = :white,
-                     bg_col::Symbol = :black)
+function plot_tfbdry end
 
+"""
+    plot_tfbdry!([plt], tree, [depth; start, node_color, line_color, background_color])
+
+Given a tree, output a visual representation of the leaf nodes on top of the current plot
+object `plt`, user will have the option to start the node count of each level with 0 or 1.
+
+# Arguments
+- `plt::Plots.Plot`: Plot object.
+- `tree::BitVector`: Tree for plotting the leaf nodes. Comes in the form of a `BitVector`.
+- `depth::Integer`: (Default: `log2(length(tree)+1) |> Int`) Maximum depth to be displayed.
+
+# Keyword Arguments
+- `start::Integer`: (Default: `0`) Whether to zero-index or one-index the root of the tree.
+- `node_color::Symbol`: (Default: `:black`) Color of the leaf nodes.
+- `line_color::Symbol`: (Default: `:black`) Color of lines in plot.
+- `background_color::Symbol`: (Default: `:white`) Color of background.
+
+# Returns
+`::Plots.Plot`: Plot object with the visual representation of the leaf nodes.
+
+# Examples
+```julia
+using Wavelets, WaveletsExt
+
+# Build a tree using Wavelets `maketree`
+tree = maketree(128, 7, :dwt)
+
+# Plot the leaf nodes
+plot_tfbdry!(tree)
+```
+
[…]
