### 2021-12-29 2b4206e
Add JOSS link to README
