### 2022-01-15 UCD4IDS/WaveletsExt.jl#42 @lostanlen
Issue templates, contributing guidelines
Hello again, JOSS reviewer here.

In the spirit of

>  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

it would be good to add a CONTRIBUTING.md document and make issue templates for bug reports and feature requests.

See as example: https://github.com/librosa/librosa/blob/main/CONTRIBUTING.md
