### 2022-01-15 #3937 @lostanlen
I have raised one issue regarding community guidelines (probably the last before approval)
https://github.com/UCD4IDS/WaveletsExt.jl/issues/42
