### 2022-01-16 UCD4IDS/WaveletsExt.jl#42 @zengfung
Thanks for the heads up. We will work on it now.
