### 2022-01-16 #3937 @lostanlen
Update: the issue above has been closed, thanks to @zengfung. At this stage i approve publication in JOSS.
