### 2022-01-21 #3937 @malmaud
I completed my review and recommend acceptance.

* This package is using all the latest best-practices for Julia packages (eg the Project.toml dependency system, CI based on Github actions, modular code organization under the `src` directory).
* The documentation is thorough.
* I was able to locally install the package and run the unit tests without error.
* As far as I know, there is no other native Julia package offering a similarly robust implementation of these wavelet methods.
