### 2022-01-21 #3937 @bmcfee
Thanks @lostanlen and @malmaud !

I'm proof-reading the paper now, and have the following comments / suggestions:

- line 43/44: I think this would be a bit easier to read if you changed the order of "acdwt" and "sdwt" to agree with the order of presentation in the previous sentence.
- Only your first example is numbered, the rest are unnumbered subsections.  I think they should be either all numbered, or none of them should be.

Everything else looks good to me so far.
