### 2022-01-21 1f84106
Update presentation order + subsection numbering
- Changed the order of “Autocorrelation Wavelet Transform” and “Stationary Wavelet Transform” to agree with the order of presentation in the next sentence and the example code.
- Add numbering to all subsections in “Examples”.
