### 2022-01-21 #3937 @zengfung
Thanks @bmcfee for your suggestions! I’ve updated the paper in the most recent [commit](https://github.com/UCD4IDS/WaveletsExt.jl/commit/1f841062c3537e64f8beaad2462099fcdd30d5e8) to the `master` branch.
