### 2022-01-21 #3937 @bmcfee
Great, thanks @zengfung !  Next steps for you:

1. Create and tag release of the package (unless the most recent release already addresses all reviewer comments)
2. Archive the release (eg at zenodo or figshare)
3. Report back here with the version number and DOI for the archive
