### 2022-01-21 354a342
Update code examples for paper/paper.md to follow syntax of v0.1.14
