### 2022-01-21 b795b64
Update to VERSION v0.1.15
No changes in functionalities in current patch update. Only fixes and additions are made to
follow JOSS guidelines.
