### 2022-01-21 8d12bf2
Merging updates to conform to JOSS guidelines (#46)
* Update code examples for paper/paper.md to follow syntax of v0.1.14

* Update to VERSION v0.1.15

* No changes in functionalities in current patch update. Only fixes and additions are made to
follow JOSS guidelines.

* Remove OhMyREPL and Revise from Project.toml
