### 2022-01-22 #3937 @zengfung
Hi @bmcfee, I have completed the steps you mentioned.

1. Package has been tagged to [v0.1.16](https://github.com/UCD4IDS/WaveletsExt.jl).
2. Release is archived at Zenodo with DOI: [10.5281/zenodo.5889625](https://zenodo.org/record/5889625#.YeuKUurMK3A)
