### 2022-01-23 f0ac240
Split `plot_tfbdry` code into multiple lines (#47)
* Fixes to JOSS code snippets. by splitting code into multiple lines to fit into 
letter size paper

* Update to VERSION v0.1.17
