### 2022-01-23 #3937 @zengfung
Hi @bmcfee I was just proof-reading our work for one last time, and I found that some of the code snippets contain lines of codes that are too long. I made the necessary fixes, and the updates are:

VERSION: v0.1.17
DOI: 10.5281/zenodo.5893843

I hope this isn't too late for changes. Apologies for the inconvenience caused.
